requirements.txt
xmlschema>=2.0.0
lxml>=4.9.0
requests>=2.28.0
python-dotenv>=1.0.0
schedule>=1.2.0
Flask>=3.0.0
Flask-Session>=0.5.0
msal>=1.24.0
mysql-connector-python>=8.2.0
PyMySQL>=1.1.0
urllib3>=2.0.0
gunicorn>=21.2.0
